+++ crates/store/src/server/api.rs
@@ -369,8 +369,7 @@ impl api_server::Api for StoreApi {
             nullifier_count = block.nullifiers().len(),
         );
 
-        // TODO: Why the error is swallowed here? Fix or add a comment with explanation.
-        let _ = self.state.apply_block(block).await;
+        self.state.apply_block(block).await?;
 
         Ok(Response::new(ApplyBlockResponse {}))
     }